# controller/specs/swagger.yml
...
swagger: '2.0'
info:
  version: 0.16.54
  title: Ziti Fabric
  description: OpenZiti Fabric API
  contact:
    name: OpenZiti
    url: https://openziti.discourse.group
    email: help@openziti.org
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: demo.ziti.dev
basePath: /fabric/v1
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

#######################################################################################################################
#
# Paths - Endpoints and their associated HTTP methods
#
#######################################################################################################################
paths:
...
  '/services':
    get:
      summary: List services
      description: |
        Retrieves a list of service resources; supports filtering, sorting, and pagination. Requires admin access.
      tags:
        - Service
      operationId: listServices
      parameters:
        - $ref: '#/parameters/limit'
        - $ref: '#/parameters/offset'
        - $ref: '#/parameters/filter'
      responses:
        '200':
          $ref: '#/responses/listServices'
        '401':
          $ref: '#/responses/unauthorizedResponse'
    post:
      summary: Create a service resource
      description: Create a service resource. Requires admin access.
      tags:
        - Service
      operationId: createService
      parameters:
        - name: service
          in: body
          required: true
          description: A service to create
          schema:
            $ref: '#/definitions/serviceCreate'
      responses:
        '201':
          $ref: '#/responses/createResponse'
        '400':
          $ref: '#/responses/badRequestResponse'
        '401':
          $ref: '#/responses/unauthorizedResponse'
...
  '/routers':
    get:
      summary: List routers
      description: |
        Retrieves a list of router resources; supports filtering, sorting, and pagination. Requires admin access.
      tags:
        - Router
      operationId: listRouters
      parameters:
        - $ref: '#/parameters/limit'
        - $ref: '#/parameters/offset'
        - $ref: '#/parameters/filter'
      responses:
        '200':
          $ref: '#/responses/listRouters'
        '401':
          $ref: '#/responses/unauthorizedResponse'
    post:
      summary: Create a router resource
      description: Create a router resource. Requires admin access.
      tags:
        - Router
      operationId: createRouter
      parameters:
        - name: router
          in: body
          required: true
          description: A router to create
          schema:
            $ref: '#/definitions/routerCreate'
      responses:
        '201':
          $ref: '#/responses/createResponse'
        '400':
          $ref: '#/responses/badRequestResponse'
        '401':
          $ref: '#/responses/unauthorizedResponse'
  '/routers/{id}':
    parameters:
      - $ref: '#/parameters/id'
    get:
      summary: Retrieves a single router
      description: Retrieves a single router by id. Requires admin access.
      tags:
        - Router
      operationId: detailRouter
      responses:
        '200':
          $ref: '#/responses/detailRouter'
        '404':
          $ref: '#/responses/notFoundResponse'
        '401':
          $ref: '#/responses/unauthorizedResponse'
    put:
      summary: Update all fields on a router
      description: Update all fields on a router by id. Requires admin access.
      tags:
        - Router
      operationId: updateRouter
      parameters:
        - name: router
          in: body
          required: true
          description: A router update object
          schema:
            $ref: '#/definitions/routerUpdate'
      responses:
        '200':
          $ref: '#/responses/updateResponse'
        '400':
          $ref: '#/responses/badRequestResponse'
        '404':
          $ref: '#/responses/notFoundResponse'
        '401':
          $ref: '#/responses/unauthorizedResponse'
    patch:
      summary: Update the supplied fields on a router
      description: Update the supplied fields on a router. Requires admin access.
      tags:
        - Router
      operationId: patchRouter
      parameters:
        - name: router
          in: body
          required: true
          description: A router patch object
          schema:
            $ref: '#/definitions/routerPatch'
      responses:
        '200':
          $ref: '#/responses/patchResponse'
        '400':
          $ref: '#/responses/badRequestResponse'
        '404':
          $ref: '#/responses/notFoundResponse'
        '401':
          $ref: '#/responses/unauthorizedResponse'
    delete:
      summary: Delete a router
      description: Delete a router by id. Requires admin access.
      tags:
        - Router
      operationId: deleteRouter
      responses:
        '200':
          $ref: '#/responses/deleteResponse'
        '400':
          $ref: '#/responses/badRequestResponse'
        '401':
          $ref: '#/responses/unauthorizedResponse'
        '409':
          $ref: '#/responses/cannotDeleteReferencedResourceResponse'
...
parameters:
  id:
    name: id
    required: true
    type: string
    in: path
    description: The id of the requested resource
  limit:
    name: limit
    type: integer
    in: query
  offset:
    name: offset
    type: integer
    in: query
  filter:
    name: filter
    type: string
    in: query

#######################################################################################################################
#
# Responses - Envelope Wrapped Definitions
#
#######################################################################################################################
responses:
...
  badRequestResponse:
    description: The supplied request contains invalid fields or could not be parsed (json and non-json bodies). The error's code, message, and cause fields can be inspected for further information
    schema:
      $ref: '#/definitions/apiErrorEnvelope'
    examples:
      'application/json':
        error:
          args:
            urlVars: {}
          cause:
            field: '(root)'
            type: required
            value:
              fooField: abc
              fooField2: def
            message: '(root): fooField3 is required'
            details:
              context: '(root)'
              field: '(root)'
              property: fooField3
          causeMessage: schema validation failed
          code: COULD_NOT_VALIDATE
          message: The supplied request contains an invalid document
          requestId: ac6766d6-3a09-44b3-8d8a-1b541d97fdd9
        meta:
          apiEnrollmentVersion: 0.0.1
          apiVersion: 0.0.1
  unauthorizedResponse:
    description: The currently supplied session does not have the correct access rights to request this resource
    schema:
      $ref: '#/definitions/apiErrorEnvelope'
    examples:
      'application/json':
        error:
          args:
            urlVars: {}
          cause: ''
          causeMessage: ''
          code: UNAUTHORIZED
          message: The request could not be completed. The session is not authorized or the
            credentials are invalid
          requestId: 0bfe7a04-9229-4b7a-812c-9eb3cc0eac0f
        meta:
          apiEnrollmentVersion: 0.0.1
          apiVersion: 0.0.1
...
  notFoundResponse:
    description: The requested resource does not exist
    schema:
      $ref: '#/definitions/apiErrorEnvelope'
    examples:
      'application/json':
        error:
          args:
            urlVars:
              id: 71a3000f-7dda-491a-9b90-a19f4ee6c406
          cause:
          causeMessage: ''
          code: NOT_FOUND
          message: The resource requested was not found or is no longer available
          requestId: 270908d6-f2ef-4577-b973-67bec18ae376
        meta:
          apiEnrollmentVersion: 0.0.1
          apiVersion: 0.0.1
  cannotDeleteReferencedResourceResponse:
    description: The resource requested to be removed/altered cannot be as it is referenced by another object.
    schema:
      $ref: '#/definitions/apiErrorEnvelope'
    examples:
      'application/json':
        error:
          args:
            urlVars:
              id: 71a3000f-7dda-491a-9b90-a19f4ee6c406
          causeMessage: 'referenced by /some-resource/05f4f710-c155-4a74-86d5-77558eb9cb42'
          code: CONFLICT_CANNOT_MODIFY_REFERENCED
          message: The resource cannot be deleted/modified. Remove all referencing resources first.
          requestId: 270908d6-f2ef-4577-b973-67bec18ae376
        meta:
          apiEnrollmentVersion: 0.0.1
          apiVersion: 0.0.1
...
  deleteResponse:
    description: The delete request was successful and the resource has been removed
    schema:
      $ref: '#/definitions/empty'
  updateResponse:
    description: The update request was successful and the resource has been altered
    schema:
      $ref: '#/definitions/empty'
  patchResponse:
    description: The patch request was successful and the resource has been altered
    schema:
      $ref: '#/definitions/empty'
  createResponse:
    description: The create request was successful and the resource has been added at the following location
    schema:
      $ref: '#/definitions/createEnvelope'
  ###################################################################
  # Services
  ##################################################################
  listServices:
    description: A list of services
    schema:
      $ref: '#/definitions/listServicesEnvelope'
...
  listRouters:
    description: A list of routers
    schema:
      $ref: '#/definitions/listRoutersEnvelope'
  detailRouter:
    description: A single router
    schema:
      $ref: '#/definitions/detailRouterEnvelope'

  ###################################################################
  # Terminators
  ##################################################################
...
definitions:
...
  pagination:
    type: object
    required:
      - limit
      - offset
      - totalCount
    properties:
      limit:
        type: number
        format: int64
      offset:
        type: number
        format: int64
      totalCount:
        type: number
        format: int64
  empty:
    type: object
    required:
      - meta
      - data
    properties:
      meta:
        $ref: '#/definitions/meta'
      data:
        type: object
        example: {}
  meta:
    type: object
    properties:
      filterableFields:
        type: array
        items:
          type: string
        x-omitempty: true
      pagination:
        $ref: '#/definitions/pagination'
      apiEnrollmentVersion:
        type: string
      apiVersion:
        type: string
  createEnvelope:
    type: object
    properties:
      meta:
        $ref: '#/definitions/meta'
      data:
        $ref: '#/definitions/createLocation'
  createLocation:
    type: object
    properties:
      _links:
        $ref: '#/definitions/links'
      id:
        type: string
  baseEntity:
    description: Fields shared by all Edge API entities
    type: object
    required:
      - id
      - createdAt
      - updatedAt
      - _links
      - tags
    properties:
      id:
        type: string
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
      _links:
        $ref: '#/definitions/links'
      tags:
        $ref: '#/definitions/tags'
  link:
    description: A link to another resource
    type: object
    required:
      - href
    properties:
      href:
        type: string
        format: uri
      method:
        type: string
      comment:
        type: string
  links:
    description: A map of named links
    type: object
    x-omitempty: false
    additionalProperties:
      $ref: '#/definitions/link'
...
  tags:
    description: 'A map of user defined fields and values. The values are limited to the following types/values: null, string, boolean'
    allOf:
      - $ref: '#/definitions/subTags'
    x-nullable: true
  subTags:
    type: object
    additionalProperties: {
      type: object
    }
  versionInfo:
    description: 'Application build information'
    type: object
    properties:
      version:
        type: string
      revision:
        type: string
      buildDate:
        type: string
      os:
        type: string
      arch:
        type: string
...
  apiErrorEnvelope:
    type: object
    required:
      - meta
      - error
    properties:
      meta:
        $ref: '#/definitions/meta'
      error:
        $ref: '#/definitions/apiError'
  apiError:
    type: object
    properties:
      args:
        $ref: '#/definitions/apiErrorArgs'
      data:
        type: object
        additionalProperties: true
      causeMessage:
        type: string
      code:
        type: string
      message:
        type: string
      requestId:
        type: string
      cause:
        $ref: '#/definitions/apiErrorCause'
  apiFieldError:
    type: object
    properties:
      reason:
        type: string
      field:
        type: string
      value:
        type: string
  apiErrorCause:
    allOf:
      - $ref: '#/definitions/apiFieldError'
      - $ref: '#/definitions/apiError'
  apiErrorArgs:
    type: object
    properties:
      urlVars:
        properties: {}
        additionalProperties:
          type: string
        type: object

  ###################################################################
  # Services
  ##################################################################
  listServicesEnvelope:
    type: object
    required:
      - meta
      - data
    properties:
      meta:
        $ref: '#/definitions/meta'
      data:
        $ref: '#/definitions/serviceList'
...
  serviceList:
    type: array
    items:
      $ref: '#/definitions/serviceDetail'
  serviceDetail:
    type: object
    allOf:
      - $ref: '#/definitions/baseEntity'
      - type: object
        required:
          - name
          - terminatorStrategy
        properties:
          name:
            type: string
          terminatorStrategy:
            type: string
  serviceCreate:
    type: object
    required:
      - name
    properties:
      name:
        type: string
      terminatorStrategy:
        type: string
      tags:
        $ref: '#/definitions/tags'
...
  listRoutersEnvelope:
    type: object
    required:
      - meta
      - data
    properties:
      meta:
        $ref: '#/definitions/meta'
      data:
        $ref: '#/definitions/routerList'
  detailRouterEnvelope:
    type: object
    required:
      - meta
      - data
    properties:
      meta:
        $ref: '#/definitions/meta'
      data:
        $ref: '#/definitions/routerDetail'
  routerList:
    type: array
    items:
      $ref: '#/definitions/routerDetail'
  routerDetail:
    type: object
    allOf:
      - $ref: '#/definitions/baseEntity'
      - type: object
        required:
          - name
          - fingerprint
          - connected
          - cost
          - noTraversal
          - disabled
        properties:
          name:
            type: string
          fingerprint:
            type: string
          connected:
            type: boolean
          cost:
            type: integer
            minimum: 0
            maximum: 65535
          noTraversal:
            type: boolean
          disabled:
            type: boolean
          listenerAddresses:
            type: array
            items:
              $ref: '#/definitions/routerListener'
          versionInfo:
            $ref: '#/definitions/versionInfo'
  routerListener:
    type: object
    required:
      - address
      - protocol
    properties:
      address:
        type: string
      protocol:
        type: string
  routerCreate:
    type: object
    required:
      - id
      - name
      - cost
      - noTraversal
    properties:
      id:
        type: string
      name:
        type: string
      fingerprint:
        type: string
        x-nullable: true
      cost:
        type: integer
        minimum: 0
        maximum: 65535
      noTraversal:
        type: boolean
      disabled:
        type: boolean
        x-nullable: true
      tags:
        $ref: '#/definitions/tags'
  routerUpdate:
    type: object
    required:
      - name
      - fingerprint
      - cost
      - noTraversal
    properties:
      name:
        type: string
      fingerprint:
        type: string
      cost:
        type: integer
        minimum: 0
        maximum: 65535
      noTraversal:
        type: boolean
      disabled:
        type: boolean
        x-nullable: true
      tags:
        $ref: '#/definitions/tags'
  routerPatch:
    type: object
    properties:
      name:
        type: string
      fingerprint:
        type: string
        x-nullable: true
      cost:
        type: integer
        minimum: 0
        maximum: 65535
        x-nullable: true
      noTraversal:
        type: boolean
        x-nullable: true
      disabled:
        type: boolean
        x-nullable: true
      tags:
        $ref: '#/definitions/tags'

  ###################################################################
  # Terminators
  ##################################################################
...